requirements.txt
flask
pillow
rembg
opencv-python-headless
numpy
gunicorn
